feat: create sessionService to extract login logic
