+++ src/app/components/user-footer/user-footer.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { User } from 'src/app/models/user.model';
 import { faSignOutAlt } from '@fortawesome/free-solid-svg-icons';
-import { OldSpotifyService } from 'src/app/services/old-spotify.service';
 import { FaIconComponent } from '@fortawesome/angular-fontawesome';
+import { SessionService } from 'src/app/services/session.service';
 
 @Component({
   selector: 'app-user-footer',
@@ -15,11 +15,11 @@ export class UserFooterComponent {
   user: User = null;
   exitIcon = faSignOutAlt;
 
-  constructor(private oldSpotifyService: OldSpotifyService) {
-    this.user = this.oldSpotifyService.user;
+  constructor(private sessionService: SessionService) {
+    this.user = this.sessionService.user;
   }
 
   logout() {
-    this.oldSpotifyService.logout();
+    this.sessionService.logout();
   }
 }
