+++ src/app/services/old-spotify.service.ts
@@ -1,15 +1,12 @@
 import Spotify from 'spotify-web-api-js';
-import { Router } from '@angular/router';
-import { Injectable, inject } from '@angular/core';
-import { User } from 'src/app/models/user.model';
+import { Injectable } from '@angular/core';
 import { Music } from 'src/app/models/music.model';
+import { SessionService } from './session.service';
 import { Artist } from 'src/app/models/artist.model';
 import { Playlist } from 'src/app/models/playlist.model';
-import { SpotifyConfiguration } from 'src/environments/environment';
 import {
   convertSportifyPlaylistToCustomPlaylist,
   convertSportifyPlaylistTracksToCustomPlaylist,
-  convertSportifyUserToCustomUser,
   convertSpotifyArtistToCustomArtist,
   convertSpotifyTrackToCustomMusic,
 } from 'src/app/common/spotifyHelper';
@@ -21,62 +18,12 @@ import {
 export class OldSpotifyService {
   spotifyAPI: Spotify.SpotifyWebApiJs = null;
 
-  user: User;
-
-  constructor(private router: Router) {
+  constructor(private sessionService: SessionService) {
     this.spotifyAPI = new Spotify();
   }
 
-  async startUser() {
-    if (!!this.user) {
-      return true;
-    }
-
-    const token = localStorage.getItem('token');
-    if (!token) {
-      return false;
-    }
-
-    try {
-      this.setAccessToken(token);
-      await this.getSpotifyUser();
-      return !!this.user;
-    } catch (error) {
-      return false;
-    }
-  }
-
-  getLoginUrl() {
-    const authEndpoint = `${SpotifyConfiguration.authEndpoint}?`;
-    const clientId = `client_id=${SpotifyConfiguration.clientId}&`;
-    const redirectUri = `redirect_uri=${SpotifyConfiguration.redirectUrl}&`;
-    const scopes = `scope=${SpotifyConfiguration.scopes.join('%20')}&`;
-    const responseType = 'response_type=token&show_dialog=true';
-
-    return authEndpoint + clientId + redirectUri + scopes + responseType;
-  }
-
-  getTokenFromUrlCallback() {
-    if (!window.location.hash) {
-      return '';
-    }
-
-    const params = window.location.href.substring(1).split('&');
-    return params[0].split('=')[1];
-  }
-
-  async getSpotifyUser() {
-    const userInfo = await this.spotifyAPI.getMe();
-    this.user = convertSportifyUserToCustomUser(userInfo);
-  }
-
-  setAccessToken(token: string) {
-    this.spotifyAPI.setAccessToken(token);
-    localStorage.setItem('token', token);
-  }
-
   async getUserPlaylists(offset = 0, limit = 50): Promise<Playlist[]> {
-    const playlists = await this.spotifyAPI.getUserPlaylists(this.user.id, { offset, limit });
+    const playlists = await this.spotifyAPI.getUserPlaylists(this.sessionService.user.id, { offset, limit });
 
     // Note: mesma coisa que fazer o "map completo"
     return playlists.items.map(convertSportifyPlaylistToCustomPlaylist);
@@ -125,10 +72,5 @@ export class OldSpotifyService {
 
   async nextMusic() {
     await this.spotifyAPI.skipToNext();
-  }
-
-  logout() {
-    localStorage.clear();
-    this.router.navigate(['/login']);
   }
 }
